backend/requirements.txt

bcrypt==3.2.2
fastapi==0.128.0
passlib==1.7.4
psycopg2==2.9.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
SQLAlchemy==2.0.46
uvicorn==0.40.0


